+++ main.go
@@ -4,6 +4,8 @@ import (
 	"flag"
 	"strconv"
 
+	"github.com/lokks307/adr-boilerplate/action"
+	"github.com/lokks307/adr-boilerplate/domain"
 	"github.com/lokks307/adr-boilerplate/env"
 	"github.com/sirupsen/logrus"
 )
@@ -23,6 +25,16 @@ func main() {
 	echoRouter.Init()
 	if *debugFlag {
 		logrus.SetLevel(logrus.TraceLevel)
+	}
+
+	if err := action.PreLoad(); err != nil {
+		logrus.Error("action preload err=", err)
+		return
+	}
+
+	if err := domain.DBLoad(); err != nil {
+		logrus.Error("domain LoadDB err=", err)
+		return
 	}
 
 	echoRouter.Run(":" + strconv.Itoa(*portNum))
